requirements.txt
flask==3.0.3
youtube-transcript-api==0.6.2
scikit-learn==1.4.2
openai==1.30.1
python-dotenv==1.0.1
yt-dlp==2024.8.6
reportlab==4.2.0
gunicorn==22.0.0
flask-limiter==3.7.0
sentence-transformers==2.7.0
transformers==4.41.2
torch==2.3.1
httpx==0.27.0